scheduler: pull tenant-querier assignments & tests into separate file (#8131)
